Add back empty userqtl b/c default is not empty!
This was the cause of 3 observed issues after my refactoring:
1) Runtime was increasing
2) Bubble trajectory was slightly different despite no changes
   to case parameters or velocity/LS equations.
3) For some combinations of parameters (e.g. after trying to
   scale the domain up by 2x) the velocity solve would fail.

This is because with an omitted userqtl, mkuserfile generates
a default userqtl that is NOT EMPTY -- it calls userqtl_scig
(which documentation describes as "thermal divergence model
for a single component ideal gas") adding a term to the velocity
equation that reads in the temperature (zeroth) passive scalar
field.

But in level set cases we use that passive scalar for the CLS
field, not temperature, so this resulted in a garbage term in
the velocity equation which increases solve time (1), changes
the bubble evolution (2), and can cause the velocity solver to
throw up completely (3).