Double domain height & remove side wall BCs
Even doubling the domain does not fix the side wall sticking. We
don't need side walls per se (as they are nonessential to adding
the source term, which we solved with the jump-periodic BC)

Even doubling the domain does not fix the side wall sticking. We
don't need side walls per se (as they are nonessential to adding
the source term, which we solved with the jump-periodic BC).